Major changes:
- Numerous bugfixes (see below).
- Completed non-ListBox-specific drag and drop support.
- Completely rewrote ListBox.

General Details:
- Grooming.
- Changed Wnd::Render() to a void function, now that ListBox no longer
needs special treatment.
- Broke out part of App::Remove() into App::WndDying(), since some removals
are done when a window is not dying, and so need to be handled differently.
- Implemented minimum drag times and distances for regular drags, and of
course for drag and drop drags as well.
- Moved MatchesOrContains() out of the App implementation, making it
generally available.
- Added full initialization of scalars in Font and related classes' ctors,
correcting Valgrind errors resulting from the use of uninitialized values.
- Corrected a buffer overreach at the end of Font::DetermineLines() when
processing an empty string.
- Added the Layout::DetachAndResetChildren() and code that calls it in
Wnd::RemoveLayout(); this allows the pair of calls SetLayout(*),
RemoveLayout() to be idempotent.
- Corrected a problem with the way Layout placed children that had exactly
one of vertical or horizontal alignments defined.
- Added output stream operators for Pt and Rect to aid instrumentation.
- Added Wnd::RelativeUpperLeft() and Wnd::RelativeLowerRight() methods that
return m_upperleft and m_lowerright directly.

ListBox Details:
- Removed the Row concepts of subrows and indentation, leaving them for
subclasses to implement as needed.
- Made Row behave as a proper Wnd subclass, including having its Control
children attached to it directly, and using Layout to position them.
- Removed Cell class.
- Changed all ListBoxStyle-typed alignment values to Alignment types; only
the flags in m_style are still LB_* alignment values.
- Renamed Delete() to Erase(), and changed the semantics from deleting Rows
to simply removing them and returning a pointer.
- Removed the concept of a default row height.
- Changed the awkward naming for the functions that affect row
highlighting.
- Fixed column headers.
- Changed storage of Rows from boost::shared_ptrs to plain pointers, not
that the drag and drop mechanism works better.