This commit represents a major repair of Texture.  Textures that were not power-of-two in each dimenstion were supposed to be placed into a proper power-of-two texture, so that the original image would be preserved, unaltered, for later retreival via texture coords.  This functionality did not work at all.  This rewrite fixes this, and an offsetting bug in SubTexture, so that now pixel accuracy is achieved with or without mipmaps and with or without the original image having power-of-two dimensions.
In the process of doing this rewrite, it was possible to remove the dependence
on ILUT, since IL's image loading is all that was really needed.  This in turn
means that all Allegro-related issues are resolved, eliminating the need for
the "Allegro hack" and generation of GG/Config.h.

The Texture API was cleaned up a bit as well.  OrthoBlit has no int-based
overloads, and has lost the useless enter_2d_mode flag.  Also, the
simplification that the number of bytes per pixel maps one-one onto the GL
storage format was removed; the Init() methods now require the GL storage
format and data type.