This commit adds Unicode support to GG.  Specific changes:
- Added the small utfcpp library to source tree under GG/utf8.

- Added UnicodeCharset class, which defines a single Unicode character set
("script" in Unicode parlance), and defined several functions that provide and
query such character sets.

- Changed Font's rendering code to treat all characters as UTF-8 encoded, and
changed the way character sets are defined in order to support an arbitrary
number of UnicodeCharsets in a single Font object.  These changes together go
most of the way to making GG fully Unicode-capable.

- Changed the Wnd::KeyPress() and Wnd::KeyRelease() functions, and
WndEvent::KeyPress, and WndEvent::KeyRelease events, to optionally contain a
Unicode code point.  If nonzero, this code point is used to generate UTF-8
encoded sequences that can be used in textual Controls.

- Updated the two reference drivers to suppy code points when doing keyboard
input if possible (the SDL driver), or 0 otherwise (the OIS input driver).

- Added an overload of TextControl::Insert() to accept std::strings, in order
to facilitate insertion of multibyte code points.