- Added a new strong-typedef macro GG_STRONG_SIZE_TYPEDEF that makes std::size_t-based string typedefs.  It does not provide interoperability with double.
- Added two new GG_STRONG_SIZE_TYPEDEF-based types, StrSize and CharSize.
They respectively represent the character position within a std::string and
the code point index of a glyph of renderable text.  These were needed to
untangle the mess that has existed in Edit and MultiEdit since the recent
addition of Unicode support.

- Changed the members of TextElement slightly to make them more uniform.  Now
the "text" member represents the original text of the element, and
FormattingTag::tag_name represents the tag name text only.  Previously, the
meaning of "text" was overloaded to mean tag text or full element text
depending on whether it was found in a TextElement or FormattingTag struct.

- Lots of header file grooming and documentation improvements.